Bug fix: Incoming overflow into a pattern causes an error when it is longer than the pattern itself.
This error could occur when handling incoming overflow for tracks not included in the pattern. Now that SongOptimizer is splitting patterns into smaller chunks, it became more likely to occur.

This situation is now handled properly: the first pattern.length # of samples are added to the pattern, and the remaining samples are added to the outgoing overflow for the next pattern to handle.