Adding SongSplitter class.
Takes a Song, and spawns a new Song for each Track. This way, to generate split output, you just use the normal combined generation code for each track-song. This removes the need to have separate logic in the core engine for generating combined and split output. This commit should allow the generation code to be simplified in future commits.

bin/beats has been updated to use this new class for generating split output. It is also using SongOptimizer on split output, so split generation is now faster than before. (The old code is being kept temporarily for profiling purposes).