Switching YAML engine from Syck -> Psych
Syck is not supported in Ruby 2.0+ without jumping through hoops,
so giving in to the inevitable. The reason for continuing to use
Syck previously was to allow support for starting tracks with the
"|" character for backwards compatibility. This is not valid YAML,
and was seemingly only allowed due to a bug in Syck.