Tracks can be defined using 'composite' sounds
That is, a track can now be defined like this:

- [sound1, sound2]: X...X...X...X...

Which is a short-hand way of expressing two tracks that have the
same rhythm, but different sounds:

- sound1: X...X...X...X...
- sound2: X...X...X...X...

This can be useful if, for example, you want a snare sound to be
created by combining two different sounds together.

As part of implementing this, kit parsing now happens after pattern
parsing, so that it can be aware of new tracks that are added as a
result of parsing tracks with composite sounds.