Kits can contain composite sounds
For example, a Kit can contain a sound that is defined like this:

- combo_snare: [clap.wav, 808_snare.wav]

When used in a track, the track will get expaned to two tracks with
the same rhythm, one for each sound. For example, this track:

- combo_snare: X...X...X...X...

Will get expanded to:

- combo_snare-clap:      X...X...X...X...
- combo_snare-808_snare: X...X...X...X...

This commit adds tests for the main happy path. Additional tests for
other scenarios should be added in the future.