Fixing bug if composite sound in kit overridden by non-composite sound
Normally, if a sound is defined in the Kit more than once, the final
definition should be the winner and override the other definitions.
However, if a composite sound was defined before a non-composite sound,
the composite sound would not be overridden, due to a bug. For example
in this kit:

Kit:
  - sound:  [sound1.wav, sound2.wav]
  - sound: sound3.wav

`sound` should be bound to sound3.wav, not [sound1.wav, sound2.wav].
However, if this first definition was a composite sound, this was
not the case.