Refactoring toward not lower casing pattern names in erorr messages
This commit should not change behavior, but should be an incremental
step toward showing pattern names with the correct capitalization in
error messages. It's a step forward because the song header can now
be extracted without doing a blanket downcasing on all of the top
level hash keys. The pattern names are still downcased, but for
proper integration with the "Flow" section, but this is a step in
the right direction.