Better error message if flow repeat count is not a String
For example, if the count is missing the "x" prefix (i.e. is just a
raw number), or if the count is some random data type such as an Array.

This change also incidentally allows making the validation logic more
clear and simple. Relatedly, the check that the repeat count is greater
than 0 can be removed, since that condition would never be reached due
the regular expression check failing upstream (both in the old code and
the new code).