Adding additional tests that multiple copies of song components are valid
This is syntactically valid YAML, but since a hash can only have a single
value for a given key, only one should "win". In practice, this means the
final definition of a given key.

Previously there were tests that a song could have multiple song header
sections, but this expands the tests to also add tests for multiple tempo,
flow, and kit sections.