Using Unix line endings, instead of Windows line endings
That is, `\n` line endings, instead of `\r\n`

For some reason, this seems to be the only file in the repo that was
using this type of line ending. It seems to date back to the initial
commit(!), but I'm not sure how I never noticed this until now.