Removing unnecessary wrapping
There's no reason to use a `WaveFile::Buffer` object to wrap the
outgoing overflow from generating a pattern's composited sample
data. That wrapping was not actually used for anything. It was
also confusing that the wrapping `WaveFile::Buffer` was taking a
hash containing multiple sample data arrays, instead of the
expected single array. The only reason that even worked is that
`WaveFile::Buffer` doesn't validate the types of the values it is
constructed with.