Improving clarity of early exit condition
It seems more clear to state that if the track has no steps, then
there is no sample data to generate. Although if a track has no
steps then `trigger_step_lengths` will also happen to be `[0]`,
that seems like a less clear condition to check for. While reading
this code before making this commit it took me a few minutes to
figure out what the original condition was there for.

The previously present "Why throw away overflow?" code comment
seems to have been the result of confusion by my past self at some
point. Since the track has no steps and therefore no sounds
triggered, there is no sample data that exists to overflow out of
the track. I'm guessing I was confusing this with the generation
of pattern sample data, in which there might be incoming overflow
from a previously generated pattern that needs to be accounted
for. However, that doesn't happen when generating the sample data
for a track.