Re-enabling commented out test
I'm not sure why this was commented out. The commit in which the
commenting occured (97ab82b1f751506ff804ee5c57595291f45664ec)
didn't give any info about why this was done.

Besides being uncommented the test also needed to be modified so
that the expected generated sample data is correct.