More explicit kit item names for tests
This hopefully makes the kit item names more clear. For example,
"S_LONGER" seems a little more explanatory than "SL".

The "TE" item, which I believe was short for "T_EXACT", is being
renamed "T" for consistency because it has the same sample count
as "S". I don't know what "T" is supposed to mean exactly, other
than being the next letter after "S". (I think "S" was intended
to be short for "sound").

For now I'm keeping the short variable names that are used in the
actual tests, instead of renaming them to match the kit item
names. The abbreviations seem convenient to make the test
assertions a little easier to read, and the names seem less
cryptic now that they can be cross-referenced with the newly
descriptive kit items names.

The one exception to "no changes to variable names" is that the
`te` variable is being renamed to `t`, to match the new kit item
name.