Re-enabling commented out test
This appears to have been commented out since the original commit
this file was added (8e2258406aff82fdd37358e437716758ea5c3fe8). I
don't know why it was commented out, but it seems like a valid
thing to test. The expected number of tracks was incorrect though,
so that has also been changed to make the test pass correctly.