Removing `MockAudioEngine` class used by tests
As a result, tests can now test the actual `AudioEngine` class,
instead of the "close but not quite" `MockAudioEngine` subclass.

The final remaining use of `MockAudioEngine` was providing a way
to manually set the sample length of a single step, which makes
the tests simpler to implement. However, that can also be done by
setting an appropriate tempo on the relevant `Song` instance.
Although calculating this tempo adds some complexity, it seems
worth it to allow directly testing `AudioEngine`.