Simplifying tests for `AudioEngine` constructor
Most of the fixture files that were being loaded were not actually
used for anything, and there's no specific reason that the test
`Song` instances need to be loaded from fixture files instead of
being created manually. Besides succinctness, a benefit of
creating `Song` instances manually is that the relevant song
tempos can be directly indicated in the test code (instead of
indirectly in external fixture files).