Refactoring `Kit` instances used for `AudioEngine` tests
Previously, the `Kit` instances set up for these tests contained
items both for sounds used in tracks, as well as sounds that
represented expected results of audio generation output. This
mixing of concepts confused me when reading this code for the
first time in a long time, until I figured out what was going on.

The fixture `Kit` instances now only contain items for sounds that
are directly used in tracks. Objects that represent expected audio
generation output are now manually defined as needed by tests. The
hopeful result is that these tests are now easier to understand.