Refactoring sample data fixtures for `AudioEngine` tests
This takes advantage of the changes in commit
41ca1f98b8d49aa1c5e576cb620e848a564b64b2 to hopefully make these
fixtures easier to work with.

First, the samples for each sound now share the same number of
digits (1 digit for "shorter_sound", 2 digits for "sound", 3
digits for "longer_sound", etc). When reading that a calculated
sample should have an expected value of `(0 + 2 + 200) / 3`, it is
hopefully now easier to see that this is the result of combining a
sample from "shorter_sound" with a sample from "longer_sound", but
not "sound".

Second, this standardizes the sample values such that the most
significant digit of each sample frame's first sample is also the
sample frame's (1-based) index in the sample frame array. For
example, `[3, -6]`, `[30, -60]`, and `[300, -600]` are all the 3rd
samples of their respective arrays.