Faster audio generation for repeated patterns
Audio generation occurs one pattern at a time. When generating
the audio for a pattern, the incoming overflow from the previous
pattern is checked to see if the combination of pattern+incoming
overflow has been seen before. If so, the resulting composited
sample data is pulled from a cache, instead of recomputed from
scratch.

To perform this check, a hash code is computed for the object
containing the incoming overflow. However, computing this hash
code can take a nontrivial amount of time. If a song contains a
pattern repeated many times, the total time taken to repeatedly
compute this hash code can add up to be a large percentage of the
total runtime.

This commit causes the hash code to be cached as well, so that the
hash code will only ever be calculated once for each pattern and
incoming overflow combination. This sometimes results in large
performance improvements.

The amount of performance improvement depends on the amount of
repeated content in the song, i.e. how often the hash code can be
pulled from cache instead of manually computed. The best case will
generally be a song that contains nothing but repeated content.
For example, the song below (using the Casio SA-20 sound kit) has
the following before/after performance (average of 10 runs), using
Ruby 2.7.2 on a 2016 MacBook Pro:

Song:
  Flow:
    - Verse: x1000

Verse:
  - bass.wav:  X...
  - snare.wav: ..X.

Before: 0.7693229 seconds
After:  0.0973544 seconds
Delta:  0.6719685 seconds faster, a 7.9x improvement

(Times are based on the duration shown in output message shown
when Beats exits).

However, if the same pattern as above is repeated 10 times instead
of 1000, there is only a very slight improvement, and other short
songs will likely show a similar result.

Before: 0.0663693 seconds
After:  0.0615182 seconds
Delta:  0.0048511 seconds faster, a 1.08x improvement

A more “real world” 3:42 long test song, with more tracks, various
verses and choruses etc., shows a large improvement. (Song omitted
because it's a little too large for a commit message):

Before: 3.4739554 seconds
After:  0.6799661 seconds
Delta:  2.7939893 seconds faster, a 5.1x improvement

In the unlikely worst case scenario in which a song has _no_
repeated content at all, then performance should be the same as
before.